server: detach emit_server_kx() from ExpectClientHello
